+++ src/media/color/mod.rs
@@ -5,9 +5,9 @@
 // safety
 #![cfg_attr(feature = "safe_color", forbid(unsafe_code))]
 
-mod base; // trait ColorBase
-mod luminance; // Luminance
-mod namespace; // struct Color
+mod color; // Color
+mod gamma; // Gamma
+mod luminance; // Luma, Luminance
 mod rgb; // Rgb[a][8|16|F32|F64], Rgb[a]Lin[F32|F64]
 
 crate::items! { // structural access: _mods, _all, _always
@@ -17,9 +17,11 @@ crate::items! { // structural access: _mods, _all, _always
     pub use _always::*;
 
     mod _mods {
-        pub use super::{base::*, luminance::*, namespace::*, rgb::_all::*};
+        pub use super::{color::*, gamma::*, luminance::*, rgb::_all::*};
         // WIPZONE
         // pub use super::spectral::*;
+        // #[cfg(feature = "linear")]
+        // pub use super::xyz::*;
     }
     pub(super) mod _all {
         #[doc(inline)]
@@ -30,3 +32,5 @@ crate::items! { // structural access: _mods, _all, _always
 }
 // WIPZONE
 // mod spectral;
+// #[cfg(feature = "linear")]
+// mod xyz; // Xyz